requirements.txt
Django==1.9.5
django-debug-panel==0.8.2
django-debug-toolbar==1.4
djangorestframework==3.3.3
pep8==1.7.0
sqlparse==0.1.19
wsgiref==0.1.2
